Beginning to split BlockChain.hs and cleanup.

+++ src/Blockchain/VM/Memory.hs
@@ -29,7 +29,6 @@ import Foreign
 
 import qualified Blockchain.Colors as CL
 import Blockchain.ExtWord
-import Blockchain.Util
 import Blockchain.VM.OpcodePrices
 import Blockchain.VM.VMState
 import Blockchain.VM.VMM
@@ -97,7 +96,7 @@ setNewMaxSize newSize' = do
     if newSize > oldLength
       then do
         state' <- lift get
-        when (newSize > 100000000) $ liftIO $ putStrLn $ CL.red ("Warning, memory needs to grow to a huge value: " ++ show (fromIntegral newSize/1000000) ++ "MB")
+        when (newSize > 100000000) $ liftIO $ putStrLn $ CL.red ("Warning, memory needs to grow to a huge value: " ++ show (fromIntegral newSize/(1000000::Double)) ++ "MB")
         arr' <- liftIO $ V.grow (mVector $ memory state') $ fromIntegral $ (newSize+1000000)
         when (newSize > 100000000) $ liftIO $ putStrLn $ CL.red $ "clearing out memory"
         --liftIO $ forM_ [oldLength..(newSize+1000000)-1] $ \p -> V.write arr' (fromIntegral p) 0
